requirements.txt
flask
pandas
python-dotenv
google-cloud-bigquery
google-auth
sqlalchemy
psycopg2-binary
db-dtypes